% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_reveal.R
\name{apply_repeat_reveal}
\alias{apply_repeat_reveal}
\title{Title}
\usage{
apply_repeat_reveal(chunk_name, num_reveal = 10,
  seed = sample(1000:2000, num_reveal, replace = F))
}
\arguments{
\item{chunk_name}{a character string which is a code chunk name}

\item{num_reveal}{an integer the number of times to repeat the revelation}

\item{seed}{a vector with length}
}
\description{
Title
}
