% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{apply_reveal}
\alias{apply_reveal}
\title{Apply reveal in Rmarkdown file, to be used in line}
\usage{
apply_reveal(chunk_name, user_reveal = F, show_code = T, title = "",
  reg_assignment = F)
}
\arguments{
\item{chunk_name}{a character string which is a chunk name}

\item{user_reveal}{a logical for if breaks should be automatically determined or have been defined manually with "#REVEAL" message}

\item{show_code}{a logical for if the code should be displayed or not, default is TRUE}

\item{title}{a character string for a title for all the slides to display code-output evolution, default is an empty string}

\item{reg_assignment}{logical set to T if output of some object created at beginning of code chunk should be displayed}
}
\value{
a character string to be interpreted as .Rmd content
}
\description{
Apply reveal in Rmarkdown file, to be used in line
}
