% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sequential_reveal.R
\name{apply_reveal_nonsequential}
\alias{apply_reveal_nonsequential}
\title{Apply nonsequential reveal}
\usage{
apply_reveal_nonsequential(chunk_name, show_code = T, title = "",
  reg_assignment = F)
}
\arguments{
\item{chunk_name}{a character string which is a code chunk name}

\item{show_code}{a logical for whether to show the code}

\item{title}{a character string if a slide title is desired}

\item{reg_assignment}{a logical for if regular assignment is being used at the beginning of the slide - may not be working}
}
\value{
character string to be interpreted as many code chunks by Rmarkdown in Xaringan
}
\description{
Apply nonsequential reveal
}
