% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{calc_highlight}
\alias{calc_highlight}
\title{Calculate highlighting for sequential reveal}
\usage{
calc_highlight(breaks)
}
\arguments{
\item{breaks}{a vector of the lines where code breaks are needed}
}
\value{
list of vectors with highlighting at each step
}
\description{
Calculate highlighting for sequential reveal
}
\examples{
calc_highlight(c(1,5,7,10))
}
