% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sequential_reveal.R
\name{calc_highlight_non_sequential}
\alias{calc_highlight_non_sequential}
\title{A list of vectors saying where to highlight at each step}
\usage{
calc_highlight_non_sequential(parsed)
}
\arguments{
\item{parsed}{parsed code including user_non_seq column}
}
\value{
a list of integer vectors
}
\description{
A list of vectors saying where to highlight at each step
}
