% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sequential_reveal.R
\name{calc_supress}
\alias{calc_supress}
\title{Calculate a list of vectors for rows to supress at each step}
\usage{
calc_supress(parsed)
}
\arguments{
\item{parsed}{parsed code including user_non_seq column}
}
\value{
a list of integer vectors
}
\description{
Calculate a list of vectors for rows to supress at each step
}
