% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{chunk_code_get}
\alias{chunk_code_get}
\title{Code chunk as text}
\usage{
chunk_code_get(chunk_name)
}
\arguments{
\item{chunk_name}{a character string which is a chunk name}
}
\value{
the code in the chunk as a string
}
\description{
Code chunk as text
}
