% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{chunk_name_return_code_sequence}
\alias{chunk_name_return_code_sequence}
\title{Title}
\usage{
chunk_name_return_code_sequence(chunk_name, break_type = "auto",
  left_assign = F)
}
\description{
Title
}
