% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{chunk_reveal}
\alias{chunk_reveal}
\title{Title}
\usage{
chunk_reveal(chunk_name = "example_name", break_type = "auto",
  reg_assign = F,
  code_seq = chunk_name_return_code_sequence(chunk_name, break_type,
  reg_assign), num_breaks = length(code_seq), display_type = "both",
  split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
