% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_exported_functions.R
\name{chunq_reveal}
\alias{chunq_reveal}
\title{Quarto default for chunk_reveal
Convenience function; see @chunk_reveal for arguments}
\usage{
chunq_reveal(...)
}
\arguments{
\item{...}{Pass arguments through to @chunk_reveal}
}
\value{
a string object is returned will only work in 'quarto' context
}
\description{
Quarto default for chunk_reveal
Convenience function; see @chunk_reveal for arguments
}
