% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{code_base_parse}
\alias{code_base_parse}
\title{Parse code}
\usage{
code_base_parse(code)
}
\arguments{
\item{code}{code as a character string}
}
\value{
parsed code
}
\description{
Parse code
}
