% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{code_sequence_return_partial_chunks}
\alias{code_sequence_return_partial_chunks}
\title{Title}
\usage{

  code_sequence_return_partial_chunks(code_seq = chunk_name_return_code_sequence(chunk_name
  = chunk_name, break_type = break_type, reg_assign = reg_assign),
  chunk_name = "example_chunk_name", break_type = "auto",
  reg_assign = F, display_type = "both", split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
\examples{
code_sequence_return_partial_chunks(code_seq = c("ggplot(cars) +", "aes(x = speed)"))
}
