% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_create_test_code.R
\name{create_base_pipe_code}
\alias{create_base_pipe_code}
\title{Title}
\usage{
create_base_pipe_code()
}
\description{
Title
}
\examples{
create_base_pipe_code()

create_base_pipe_code() \%>\%
 code_parse()
}
