% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{create_cow_frame}
\alias{create_cow_frame}
\title{Title}
\usage{
create_cow_frame(partial_code_w_highlight, title = "flipbook mini",
  highlight_color = "plum4", font_size = 7, num_lines = 16,
  display_type = "both")
}
\arguments{
\item{partial_code_w_highlight}{}
}
\description{
Title
}
