% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_minimal_example.R
\name{create_minimal_example}
\alias{create_minimal_example}
\title{Create Rmd example}
\usage{
create_minimal_example(path, open = FALSE)
}
\arguments{
\item{path}{a character string to save out the minimal example}
}
\value{
a .Rmd
}
\description{
Create Rmd example
}
