% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_prep_sequences.R
\name{left_assign_detect}
\alias{left_assign_detect}
\title{Title}
\usage{
left_assign_detect(parsed)
}
\arguments{
\item{parsed}{}
}
\description{
Title
}
\examples{
create_left_assign_code() \%>\%
 code_parse() \%>\%
 left_assign_detect()

 create_ggplot_code() \%>\%
 code_parse() \%>\%
 left_assign_detect()

 create_code() \%>\%
 r_code_base_parse()
 code_parse() \%>\%
 left_assign_detect()
}
