% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_code.R
\docType{data}
\name{local_code}
\alias{local_code}
\alias{local_code_regular_assignment}
\title{Local Code}
\format{An object of class \code{character} of length 1.}
\usage{
local_code

local_code_regular_assignment
}
\description{
Code for examples in the flipbookr package.
}
\keyword{datasets}
