% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{parse_chunk}
\alias{parse_chunk}
\title{Parse chunk}
\usage{
parse_chunk(chunk_name)
}
\arguments{
\item{chunk_name}{a character string which is a chunk name}
}
\value{
parsed code from a code chunk in an Rmd
}
\description{
Parse chunk
}
