% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_reveal_parentheses_balanced.R
\name{parse_code}
\alias{parse_code}
\title{Parse code}
\usage{
parse_code(code)
}
\arguments{
\item{code}{code as a character string}
}
\value{
parsed code
}
\description{
Parse code
}
\examples{
parse_code(code = local_code)
parse_code(code = local_code_regular_assignment)
}
