% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_prep_sequences.R
\name{parsed_left_assign_return_partial_code}
\alias{parsed_left_assign_return_partial_code}
\title{Title}
\usage{
parsed_left_assign_return_partial_code(
  parsed,
  platform = c("xaringan", "quarto"),
  which_show_frame = 1:3,
  which_highlight_frame = 3,
  left_assign_add = "flextable::flextable()"
)
}
\arguments{
\item{left_assign_add}{}
}
\description{
Title
}
\examples{
create_left_assign_code() \%>\%
code_parse() \%>\%
parsed_left_assign_return_partial_code()
}
