% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{parsed_return_partial_code_sequence}
\alias{parsed_return_partial_code_sequence}
\title{Title}
\usage{
parsed_return_partial_code_sequence(parsed, break_type = "auto",
  which_show = parsed_calc_show(parsed = parsed, break_type =
  break_type), which_highlight = shown_lines_calc_highlight(which_show =
  which_show, break_type = break_type), left_assign = F)
}
\arguments{
\item{left_assign}{}
}
\description{
Title
}
