% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{partially_knit_chunks}
\alias{partially_knit_chunks}
\title{Title}
\usage{
partially_knit_chunks(the_code = get_partial_codes_from_chunk(chunk_name
  = chunk_name, break_type = break_type, reg_assign = reg_assign),
  chunk_name = "example_chunk_name", break_type = "auto",
  reg_assign = F, display_type = "both", split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
