% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini.R
\name{prepped_partial_plotting_code_plot_text}
\alias{prepped_partial_plotting_code_plot_text}
\title{#' Title
#'
#' @param code
#' @param upto
#' @param highlight
#' @param highlight_color
#' @param font_size
#'
#' @return
#' @export
#'
#' @examples}
\usage{
prepped_partial_plotting_code_plot_text(prepped_plotting_code,
  highlight_color = "plum4", font_size = 5, num_lines = 16)
}
\description{
#' Title
#'
#' @param code
#' @param upto
#' @param highlight
#' @param highlight_color
#' @param font_size
#'
#' @return
#' @export
#'
#' @examples
}
