% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_chunks_template_code}
\alias{return_partial_chunks_template_code}
\title{Title}
\usage{
return_partial_chunks_template_code()
}
\arguments{
\item{eval}{}
}
\description{
Title
}
\examples{
return_partial_chunks_template_code()
}
