% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_chunks_template_output}
\alias{return_partial_chunks_template_output}
\title{Title}
\usage{
return_partial_chunks_template_output()
}
\description{
Title
}
\examples{
return_partial_chunks_template_output()
}
