% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_side_by_side_code_output_chunks}
\alias{return_partial_side_by_side_code_output_chunks}
\title{Title}
\usage{

  return_partial_side_by_side_code_output_chunks(chunk_name = "a_chunk_name",
  break_type = "auto", reg_assign = F,
  code_seq = chunk_name_return_code_sequence(chunk_name = chunk_name,
  break_type = break_type, reg_assign = reg_assign),
  num_breaks = length(code_seq), split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
\examples{
return_partial_side_by_side_code_output_chunks(num_breaks = 3)
}
