% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{reveal_chunk}
\alias{reveal_chunk}
\title{Reveal Chunk}
\usage{
reveal_chunk(chunk_name, upto = 3, highlight = 1:3,
  reg_assignment = F)
}
\arguments{
\item{chunk_name}{a character string which is a chunk name}

\item{upto}{an integer indicating the line of code}

\item{highlight}{integers indicating which lines of code to highlight}

\item{reg_assignment}{logical set to T if output of some object created at beginning of code chunk should be displayed}
}
\value{
Partial code with indicators for highlighting
}
\description{
Reveal Chunk
}
