% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sequential_reveal.R
\name{reveal_chunk_nonsequential}
\alias{reveal_chunk_nonsequential}
\title{Reveal chunk content non-sequentially}
\usage{
reveal_chunk_nonsequential(chunk_name, which_supress = 2:5,
  which_highlight = 6)
}
\arguments{
\item{which_supress}{an integer vector indicating rows of code to supress}

\item{which_highlight}{an integer vector indicating rows of code to highlight}
}
\value{
redacted code with highlighting indicators
}
\description{
Reveal chunk content non-sequentially
}
