% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{reveal_code}
\alias{reveal_code}
\title{Partially reveal code}
\usage{
reveal_code(code, upto = 3, highlight = 1:3)
}
\arguments{
\item{code}{code as a character string}

\item{upto}{an integer indicating the line of code}

\item{highlight}{integers indicating which lines of code to highlight}
}
\value{
Partial code with indicators for highlighting
}
\description{
Partially reveal code
}
\examples{
reveal_code(code = local_code_regular_assignment)
reveal_code(code = local_code)
reveal_code(code = local_code, 5, 3:5)
}
