% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_sequential_reveal.R
\name{reveal_code_nonsequential}
\alias{reveal_code_nonsequential}
\title{Supress some code and highlight some code}
\usage{
reveal_code_nonsequential(code, which_supress = 2:5,
  which_highlight = 6)
}
\arguments{
\item{which_supress}{an integer vector}

\item{which_highlight}{an integer vector}
}
\description{
Supress some code and highlight some code
}
\examples{
reveal_code_nonsequential(code = local_code_non_sequential)
}
