% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_reveal_parentheses_balanced.R
\name{reveal_parsed}
\alias{reveal_parsed}
\title{Partially reveal parsed code}
\usage{
reveal_parsed(parsed, upto = 3, highlight = 1:3)
}
\arguments{
\item{parsed}{the output resulting from parsing code}

\item{upto}{a integer indicating the line of code}

\item{highlight}{integers indicating which lines of code to highlight}
}
\value{
Partial code with indicators for highlighting
}
\description{
Partially reveal parsed code
}
