% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_reveal.R
\name{reveal_sequence}
\alias{reveal_sequence}
\title{Reveal different realizations of the same code-output frames several times}
\usage{
reveal_sequence(chunk_name, num_reveal = 10, seed = sample(1000:2000,
  num_reveal, replace = F))
}
\arguments{
\item{chunk_name}{a character string which is a code chunk name}

\item{num_reveal}{an integer the number of times to repeat the revelation}

\item{seed}{a vector with length}
}
\description{
Reveal different realizations of the same code-output frames several times
}
