% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{show_and_highlight_code_frames}
\alias{show_and_highlight_code_frames}
\title{Title}
\usage{
show_and_highlight_code_frames(parsed, break_type = "auto",
  which_show = calc_lines_to_show(parsed = parsed, break_type =
  break_type), which_highlight = calc_lines_to_highlight(which_show =
  which_show, break_type = break_type), reg_assign = F)
}
\arguments{
\item{reg_assign}{}
}
\description{
Title
}
