% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{show_and_highlight_pane_classic}
\alias{show_and_highlight_pane_classic}
\title{Partially reveal parsed code}
\usage{
show_and_highlight_pane_classic(parsed, which_show = 1:3, which_highlight = 3)
}
\arguments{
\item{parsed}{the output resulting from parsing code}

\item{which_show}{default to \code{1:3}}

\item{which_highlight}{default to \code{3}}
}
\value{
Partial code with indicators for highlight
}
\description{
Partially reveal parsed code
}
\examples{
# to be written
}
