% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_exported_functions.R
\name{text_reveal}
\alias{text_reveal}
\title{Function takes character string, splits it based on delimiter,
and returns each element of the resultant vector on its own slide}
\usage{
text_reveal(text, sep = "   ", md_prefix = "#", sep_replace = "",
  slide_break = "---", class = "class: inverse, middle, center")
}
\arguments{
\item{text}{a character string to be split and delivered piece-wise to a slide}

\item{sep}{a character string to delimit the split of the input text}

\item{md_prefix}{a character string prefix to each markdown element, defaults to "#"}

\item{sep_replace}{a character string that will replace the delimiter, defaults to empty string ""}

\item{slide_break}{a character string containing slide break characters, defaults to "---" for xaringan slideshows}

\item{class}{a character string in which you can set the class, defaults to "class: inverse, middle, center"}
}
\value{
knit text to be interpreted as slides
}
\description{
Function takes character string, splits it based on delimiter,
and returns each element of the resultant vector on its own slide
}
\examples{
text_reveal("Hello world", sep = " ")
}
