% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{chunk_expand}
\alias{chunk_expand}
\title{Title}
\usage{
chunk_expand(chunk_name = "example", reveal_type = "auto",
  display_type = "both", num_breaks = 2, split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
\examples{
chunk_expand()
chunk_expand(reveal_type = "user", display_type = "code")
chunk_expand(reveal_type = "non_seq", display_type = "output")
}
