% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_chunks_template}
\alias{return_partial_chunks_template}
\title{Title}
\usage{
return_partial_chunks_template(display_type = "output",
  eval = display_type == "output", echo = display_type == "code")
}
\arguments{
\item{echo}{}
}
\description{
Title
}
\examples{
return_partial_chunks_template()
return_partial_chunks_template(display_type = "output")
}
