% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_code_or_output_chunks}
\alias{return_partial_code_or_output_chunks}
\title{Title}
\usage{
return_partial_code_or_output_chunks(chunk_name = "a_chunk_name",
  break_type = "auto", reg_assign = F,
  the_code = chunk_name_return_code_sequence(chunk_name = chunk_name,
  break_type = break_type, reg_assign = reg_assign),
  breaks = 1:length(the_code), display_type = "output")
}
\arguments{
\item{display_type}{}
}
\description{
Title
}
\examples{
return_partial_code_or_output_chunks(breaks = 1:3)
return_partial_code_or_output_chunks(display_type = "code", breaks = 1:3)
}
