% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_parse_reveal_xaringan.R
\name{return_partial_side_by_side_code_output_chunks}
\alias{return_partial_side_by_side_code_output_chunks}
\title{Title}
\usage{

  return_partial_side_by_side_code_output_chunks(chunk_name = "a_chunk_name",
  break_type = "auto", num_breaks = 2, split = 40)
}
\arguments{
\item{split}{}
}
\description{
Title
}
\examples{
return_partial_side_by_side_code_output_chunks(num_breaks = 3)
}
